+++ spec/commands/zrange_spec.rb
@@ -3,9 +3,9 @@ require 'spec_helper'
 describe '#zrange(key, start, stop [, :with_scores => true])' do
   before do
     @key = 'mock-redis-test:zrange'
+    @redises.zadd(@key, 3, 'Jefferson')
     @redises.zadd(@key, 1, 'Washington')
     @redises.zadd(@key, 2, 'Adams')
-    @redises.zadd(@key, 3, 'Jefferson')
     @redises.zadd(@key, 4, 'Madison')
   end
 
@@ -26,6 +26,16 @@ describe '#zrange(key, start, stop [, :with_scores => true])' do
 
   it 'returns the elements in order by score' do
     @redises.zrange(@key, 0, 1).should == %w[Washington Adams]
+  end
+
+  context 'when a subset of elements have the same score' do
+    before do
+      @redises.zadd(@key, 1, 'Martha')
+    end
+
+    it 'returns the elements in ascending lexicographical order' do
+      @redises.zrange(@key, 0, 1).should == %w[Martha Washington]
+    end
   end
 
   it 'returns the elements in order by score (negative indices)' do